test: admin과 user 클래스 분리
